requirements.txt
torch
torch_geometric
ogb
pyyaml
mlflow
jsonlines
google-cloud-aiplatform